convert measurements from f32 (meters) to i32 (milimeters)

+++ src/collect.rs
@@ -6,9 +6,9 @@
 #[derive(Debug, Clone, Copy)]
 pub struct MotionControlReading {
     pub valid: bool,
-    pub accel_x: f32,
-    pub accel_y: f32,
-    pub gyro_z: f32,
+    pub accel_x: i32,
+    pub accel_y: i32,
+    pub gyro_z: i32,
     pub encoder_values: [u8; 10],
     pub delta_t: u64,
 }
@@ -33,9 +33,9 @@ impl MotionControlReading {
     }
 
     pub fn from_bytes(bytes: &[u8]) -> Self {
-        let accel_x = f32::from_le_bytes(bytes[0..4].try_into().unwrap());
-        let accel_y = f32::from_le_bytes(bytes[4..8].try_into().unwrap());
-        let gyro_z = f32::from_le_bytes(bytes[8..12].try_into().unwrap());
+        let accel_x = i32::from_le_bytes(bytes[0..4].try_into().unwrap());
+        let accel_y = i32::from_le_bytes(bytes[4..8].try_into().unwrap());
+        let gyro_z = i32::from_le_bytes(bytes[8..12].try_into().unwrap());
         let encoder_values: [u8; 10] = bytes[12..22].try_into().unwrap();
         let valid = bytes[22] == 0;
         let delta_t = u64::from_le_bytes(bytes[23..31].try_into().unwrap());
